Fixed header file and unit test
- Fixed header for triangle angles function
- Fixed unit test value not being the expected

+++ PolygonChecker/triangleSolver.h
@@ -1,3 +1,3 @@
 #pragma once
 char* analyzeTriangle(int side1, int side2, int side3);
-void GetAnglesFromSides(int side1, int side2, int side3, float buffer[]);
+void GetAnglesFromSides(float side1, float side2, float side3, float buffer[]);
